- Separated classes into different files, - Removed unnecessary files
