+++ INPTPZ1/Program.cs
@@ -162,142 +162,4 @@ namespace INPTPZ1
             bmpCanvas.SetPixel(x, y, pxlColor);
         }
     }
-
-    class Polynome
-    {
-        public List<ComplexNumber> Coefficient { get; set; }
-
-        public Polynome()
-        {
-            Coefficient = new List<ComplexNumber>();
-        }
-
-        public Polynome(List<ComplexNumber> coefs)
-        {
-            Coefficient = new List<ComplexNumber>();
-            foreach (ComplexNumber element in coefs)
-            {
-                Coefficient.Add(element);
-            }
-        }
-
-        public Polynome Derivate()
-        {
-            Polynome derivation = new Polynome();
-            for (int i = 1; i < Coefficient.Count; i++)
-            {
-                derivation.Coefficient.Add(Coefficient[i].Multiply(new ComplexNumber() { RealPart = i }));
-            }
-
-            return derivation;
-        }
-
-        public ComplexNumber Evaluate(ComplexNumber complexNumber)
-        {
-            ComplexNumber evaluatedComplex = ComplexNumber.Zero;
-            for (int i = 0; i < Coefficient.Count; i++)
-            {
-                ComplexNumber coefficient = Coefficient[i];
-                ComplexNumber givenComplexNumber = complexNumber;
-                int power = i;
-
-                if (i > 0)
-                {
-                    for (int j = 0; j < power - 1; j++)
-                        givenComplexNumber = givenComplexNumber.Multiply(complexNumber);
-
-                    coefficient = coefficient.Multiply(givenComplexNumber);
-                }
-
-                evaluatedComplex = evaluatedComplex.Add(coefficient);
-            }
-
-            return evaluatedComplex;
-        }
-
-        public override string ToString()
-        {
-            string polynomeByString = "";
-            for (int i = 0; i < Coefficient.Count; i++)
-            {
-                polynomeByString += Coefficient[i];
-                if (i > 0)
-                {
-                    for (int j = 0; j < i; j++)
-                    {
-                        polynomeByString += "x";
-                    }
-                }
-                polynomeByString += " + ";
-            }
-            return polynomeByString.Substring(0, polynomeByString.Length - 3);
-        }
-    }
-
-    class ComplexNumber
-    {
-        public double RealPart { get; set; }
-        public double ImaginaryPart { get; set; }
-
-        public readonly static ComplexNumber Zero = new ComplexNumber()
-        {
-            RealPart = 0,
-            ImaginaryPart = 0
-        };
-
-        public bool isComplexZero()
-        {
-            ComplexNumber a = this;
-            if (a.RealPart == 0 && a.ImaginaryPart == 0)
-                return true;
-            else
-                return false;
-        }
-
-        public ComplexNumber Multiply(ComplexNumber b)
-        {
-            ComplexNumber a = this;
-            return new ComplexNumber()
-            {
-                RealPart = a.RealPart * b.RealPart - a.ImaginaryPart * b.ImaginaryPart,
-                ImaginaryPart = a.RealPart * b.ImaginaryPart + a.ImaginaryPart * b.RealPart
-            };
-        }
-
-        public ComplexNumber Add(ComplexNumber b)
-        {
-            ComplexNumber a = this;
-            return new ComplexNumber()
-            {
-                RealPart = a.RealPart + b.RealPart,
-                ImaginaryPart = a.ImaginaryPart + b.ImaginaryPart
-            };
-        }
-        public ComplexNumber Subtract(ComplexNumber b)
-        {
-            ComplexNumber a = this;
-            return new ComplexNumber()
-            {
-                RealPart = a.RealPart - b.RealPart,
-                ImaginaryPart = a.ImaginaryPart - b.ImaginaryPart
-            };
-        }
-
-        internal ComplexNumber Divide(ComplexNumber b)
-        {
-            ComplexNumber numerator = this.Multiply(new ComplexNumber() { RealPart = b.RealPart, ImaginaryPart = -b.ImaginaryPart });
-            double denominator = b.RealPart * b.RealPart + b.ImaginaryPart * b.ImaginaryPart;
-
-            return new ComplexNumber()
-            {
-                RealPart = numerator.RealPart / denominator,
-                ImaginaryPart = numerator.ImaginaryPart / denominator
-            };
-        }
-
-        public override string ToString()
-        {
-            return $"({RealPart} + {ImaginaryPart}i)";
-        }
-    }
 }
